update: AppDelegateを固有クラスへ書き出し
